+++ packages/popup-picker/src/popupPicker.ts
@@ -31,13 +31,13 @@ export class PopupPickerController {
 
   referenceElement?: HTMLElement;
   triggerElement?: HTMLElement;
+  options: PickerOptions & PopupOptions;
 
   private popupEl: HTMLElement;
   private focusTrap: FocusTrap;
   private positionCleanup: PositionCleanup;
 
   private closeButton: HTMLButtonElement;
-  private options: PickerOptions & PopupOptions;
   private externalEvents = new Events<PopupEvent>();
 
   constructor(pickerOptions: Partial<PickerOptions>, popupOptions: Partial<PopupOptions>) {
@@ -56,6 +56,9 @@ export class PopupPickerController {
       this.closeButton.type = 'button';
       this.closeButton.classList.add(classes.closeButton);
       this.closeButton.innerHTML = closeIcon;
+      this.closeButton.addEventListener('click', () => {
+        this.close();
+      });
       this.popupEl.appendChild(this.closeButton);
     }
 
@@ -219,6 +222,7 @@ export class PopupPickerController {
   private async setPosition() {
     this.positionCleanup?.();
       this.positionCleanup = await setPosition(
+        this,
         this.popupEl,
         this.referenceElement,
         this.options.position as Position