+++ packages/popup-picker/src/types.ts
@@ -9,6 +9,11 @@ export type FixedPosition = {
 
 export type RelativePosition = Placement | 'auto';
 export type Position = FixedPosition | RelativePosition;
+export type PositionLostStrategy = 
+  | 'close'   // close the picker immediately
+  | 'destroy' // destroy the picker permanently
+  | 'hold'    // hold the previously calculated position
+  | 'none';   // no action (the picker will jump to the top left corner of the screen if not closed manually)
 
 export type PopupOptions = {
   hideOnClickOutside: boolean;
@@ -19,4 +24,5 @@ export type PopupOptions = {
   triggerElement?: HTMLElement;
   showCloseButton?: boolean;
   className?: string;
+  onPositionLost?: PositionLostStrategy;
 }